Improve purge tasks of DbCleaner
* improve logs
* decrease the nb of SQL requests required to purge aborted builds and historical data of directories/files

+++ sonar-core/src/main/java/org/sonar/core/resource/ResourceDto.java
@@ -21,18 +21,18 @@ package org.sonar.core.resource;
 
 public final class ResourceDto {
 
-  private Integer id;
+  private Long id;
   private String name;
   private String longName;
   private Integer rootId;
   private String scope;
   private String qualifier;
 
-  public Integer getId() {
+  public Long getId() {
     return id;
   }
 
-  public ResourceDto setId(Integer id) {
+  public ResourceDto setId(Long id) {
     this.id = id;
     return this;
   }
